fix: replaced double with int

+++ src/common/AutoRobot.cpp
@@ -10,26 +10,26 @@
 
 #include "../headers/Map.h" // If put in header Qt Creator uses 100% CPU (recursive)
 
-AutoRobot::AutoRobot(double x, double y, double angle, double radius) {
+AutoRobot::AutoRobot(int x, int y, int angle, int radius) {
     this->_collider = new CircleCollider(x, y, angle, radius);
     this->_position = new Position(x, y, angle);
     this->_objectType = eot_auto_robot;
     // TODO Add collision boxes
 }
 
-double AutoRobot::GetSpeed() {
+int AutoRobot::GetSpeed() {
     return this->_speed;
 }
 
-void AutoRobot::SetSpeed(double speed) {
+void AutoRobot::SetSpeed(int speed) {
     this->_speed = speed;
 }
 
-double AutoRobot::GetRotationAngle() {
+int AutoRobot::GetRotationAngle() {
     return this->_rotationAngle;
 }
 
-void AutoRobot::SetRotationAngle(double rotationAngle) {
+void AutoRobot::SetRotationAngle(int rotationAngle) {
     this->_rotationAngle = rotationAngle;
 }
